feat: introduced API for managing secrets

+++ core/control-plane/control-plane-aggregate-services/src/main/java/org/eclipse/edc/connector/controlplane/services/secret/SecretServiceImpl.java
@@ -18,18 +18,15 @@ import org.eclipse.edc.connector.secret.spi.observe.SecretObservable;
 import org.eclipse.edc.connector.spi.service.SecretService;
 import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.spi.result.ServiceResult;
-import org.eclipse.edc.spi.result.StoreResult;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.types.domain.secret.Secret;
 
 import java.util.List;
+import java.util.Optional;
 
 public class SecretServiceImpl implements SecretService {
     private final Vault vault;
     private final SecretObservable observable;
-
-    // TODO: check is secret is referenced by an asset (privateProperties)?
-    // -> no, use Vault & observable
 
     public SecretServiceImpl(Vault vault, SecretObservable observable) {
         this.vault = vault;
@@ -38,52 +35,67 @@ public class SecretServiceImpl implements SecretService {
 
     @Override
     public Secret findById(String secretKey) {
-        var secretValue = vault.resolveSecret(secretKey);
-        // TODO: check if secretValue is null
-        return Secret.Builder.newInstance()
-                .value(secretValue)
-                .key(secretKey)
-                .build();
+        return Optional.ofNullable(vault.resolveSecret(secretKey))
+                .map(secretValue -> Secret.Builder.newInstance()
+                        .value(secretValue)
+                        .key(secretKey)
+                        .build())
+                .orElse(null);
     }
 
     @Override
     public ServiceResult<List<Secret>> search(QuerySpec query) {
-        // TODO: should this function be implemented? The vault doesn't seem to support searching for secrets
-        return null;
+        throw new UnsupportedOperationException("Query operation is not supported for secrets");
     }
 
     @Override
     public ServiceResult<Secret> create(Secret secret) {
+        var existingSecret = findById(secret.getKey());
+
+        if (existingSecret != null) {
+            return ServiceResult.conflict("Secret " + secret.getKey() + " already exist");
+        }
+
         var createResult = vault.storeSecret(secret.getKey(), secret.getValue());
         if (createResult.succeeded()) {
             observable.invokeForEach(l -> l.created(secret));
             return ServiceResult.success(secret);
         }
-        StoreResult<Secret> storeResult = StoreResult.notFound("Secret " + secret.getKey() + " not found");
-        return ServiceResult.fromFailure(storeResult);
+
+        return ServiceResult.badRequest(createResult.getFailureMessages().toString());
     }
 
     @Override
     public ServiceResult<Secret> delete(String secretKey) {
+        var existingSecret = findById(secretKey);
+
+        if (existingSecret == null) {
+            return ServiceResult.notFound("Secret " + secretKey + " not found");
+        }
+
         var deleteResult = vault.deleteSecret(secretKey);
         if (deleteResult.succeeded()) {
-            StoreResult<Secret> storeResult = StoreResult.success(null);
-            storeResult.onSuccess(a -> observable.invokeForEach(l -> l.deleted(a)));
-            return ServiceResult.from(storeResult);
+            deleteResult.onSuccess(a -> observable.invokeForEach(l -> l.deleted(secretKey)));
+            return ServiceResult.success(null);
         }
-        StoreResult<Secret> storeResult = StoreResult.notFound("Secret " + secretKey + " not found");
-        return ServiceResult.from(storeResult);
+        return ServiceResult.badRequest(deleteResult.getFailureMessages().toString());
     }
 
     @Override
-    // TODO: should we check if the secret already exist and reject update if not?
     public ServiceResult<Secret> update(Secret secret) {
+        var existingSecret = findById(secret.getKey());
+        
+        if (existingSecret == null) {
+            return ServiceResult.notFound("Secret " + secret.getKey() + " not found");
+        }
+
         var updateResult = vault.storeSecret(secret.getKey(), secret.getValue());
         if (updateResult.succeeded()) {
             observable.invokeForEach(l -> l.updated(secret));
             return ServiceResult.success(secret);
         }
-        StoreResult<Secret> storeResult = StoreResult.notFound("Secret " + secret.getKey() + " not found");
-        return ServiceResult.fromFailure(storeResult);
+        return ServiceResult.badRequest(updateResult.getFailureMessages().toString());
     }
+
 }
+
